--- ZapBuy/user/forms.py
@@ -1,23 +1,0 @@
-from django import forms
-from django.contrib.auth.models import User
-from django.contrib.auth.forms import UserCreationForm
-# from .models import Profile
-
-class UserRegisterForm(UserCreationForm):
-    email = forms.EmailField()
-
-    # def __init__(self, *args, **kwargs):
-    #     super().__init__(*args, **kwargs)
-    #     self.fields['first_name'].required = True
-    #     self.fields['last_name'].required = True
-    
-    class Meta:
-        model = User
-        fields = ['username', 'email',  'password1', 'password2']
-
-class UserUpdateForm(forms.ModelForm):
-    email = forms.EmailField()
-    class Meta:
-        model = User
-        fields = ['username', 'email']
-
