+++ django_auth/users/static/users/main.js
