+++ django_auth/users/templates/users/login.html
@@ -1,5 +1,5 @@
 <!-- login.html -->
-{% extends 'user/base.html' %}
+{% extends 'users/base.html' %}
 
 {% block content %}
 <div class="row justify-content-center">
@@ -10,11 +10,11 @@
                 <form method="post">
                     {% csrf_token %}
                     <div class="form-group">
-                        <label for="{{ form.username.id_for_label }}">Username:</label>
+                        <label for="{{ form.username.id_for_label }}">Email:</label>
                         {{ form.username }}
                     </div>
                     <div class="form-group">
-                        <label for="{{ form.password1.id_for_label }}">Enter Password:</label>
+                        <label for="{{ form.password.id_for_label }}">Enter Password:</label>
                         {{ form.password }}
                     </div>
