+++ django_auth/users/templates/users/logout.html
@@ -1,4 +1,4 @@
-{% extends 'user/base.html' %}
+{% extends 'users/base.html' %}
 {% block content %}
     <h2>You Have been Logged out</h2>
     <div class="border-top pt-3">
