+++ django_auth/users/templates/users/register.html
@@ -1,5 +1,5 @@
 <!-- registration.html -->
-{% extends 'user/base.html' %}
+{% extends 'users/base.html' %}
 
 {% block content %}
 <div class="row justify-content-center">
@@ -9,10 +9,10 @@
                 <h2 class="card-title">Registration</h2>
                 <form method="post">
                     {% csrf_token %}
-                    <div class="form-group">
+                    <!-- <div class="form-group">
                         <label for="{{ form.username.id_for_label }}">Username:</label>
                         {{ form.username }}
-                    </div>
+                    </div> -->
                     <div class="form-group">
                         <label for="{{ form.email.id_for_label }}">Email:</label>
                         {{ form.email }}
