+++ django_auth/users/views.py
@@ -1,3 +1,18 @@
-from django.shortcuts import render
+from django.shortcuts import render,redirect
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+from .forms import UserRegisterForm
 
-# Create your views here.
+
+def register(request):
+    if request.method == 'POST':
+        form = UserRegisterForm(request.POST)
+        if form.is_valid():
+            form.save()
+            email = form.cleaned_data.get('email')
+            messages.success(request, f'Account created for {email}! now login')
+            return redirect('login')
+
+    else:
+        form = UserRegisterForm()
+    return render(request, 'users/register.html', {'form':form})
